@@ -5,13 +5,13 @@ import android.content.Intent;
 import android.content.SharedPreferences;
 import android.media.MediaPlayer;
 import android.os.Bundle;
-import android.support.design.widget.NavigationView;
-import android.support.design.widget.Snackbar;
-import android.support.v4.view.GravityCompat;
-import android.support.v4.widget.DrawerLayout;
-import android.support.v7.app.ActionBarDrawerToggle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
+import com.google.android.material.navigation.NavigationView;
+import com.google.android.material.snackbar.Snackbar;
+import androidx.core.view.GravityCompat;
+import androidx.drawerlayout.widget.DrawerLayout;
+import androidx.appcompat.app.ActionBarDrawerToggle;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
@@ -39,7 +39,7 @@ import nl.rsdt.japp.jotial.net.apis.MetaApi;
 import nl.rsdt.japp.service.LocationService;
 import nl.rsdt.japp.service.cloud.data.NoticeInfo;
 import nl.rsdt.japp.service.cloud.data.UpdateInfo;
-import nl.rsdt.japp.service.cloud.messaging.JappFirebaseInstanceIdService;
+import nl.rsdt.japp.service.cloud.messaging.JappFirebaseMessagingService;
 import nl.rsdt.japp.service.cloud.messaging.MessageManager;
 import retrofit2.Call;
 import retrofit2.Callback;
@@ -119,7 +119,7 @@ public class MainActivity extends AppCompatActivity
             /**
              * Send the token to the server.
              * */
-            JappFirebaseInstanceIdService.sendToken();
+            JappFirebaseMessagingService.sendToken();
 
             /**
              * Show the user around the app via a sequence.
