@@ -1,17 +1,13 @@
 package nl.rsdt.japp.jotial.maps.locations;
 
-import android.Manifest;
 import android.app.Service;
-import android.content.Intent;
-import android.content.pm.PackageManager;
 import android.location.Location;
 import android.os.Binder;
 import android.os.Bundle;
-import android.os.IBinder;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.design.widget.Snackbar;
-import android.support.v4.app.ActivityCompat;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+
 import android.util.Log;
 
 import com.google.android.gms.common.ConnectionResult;
@@ -24,7 +20,6 @@ import com.google.android.gms.location.LocationRequest;
 import com.google.android.gms.location.LocationServices;
 import com.google.android.gms.location.LocationSettingsRequest;
 import com.google.android.gms.location.LocationSettingsResult;
-import com.google.android.gms.location.LocationSettingsStates;
 import com.google.android.gms.location.LocationSettingsStatusCodes;
 
 import java.util.ArrayList;
@@ -67,7 +62,7 @@ public abstract class LocationProviderService<B extends Binder> extends Service 
         client.connect();
     }
 
-    public void startLocationUpdates() {
+    public void startLocationUpdates() throws SecurityException{
         LocationServices.FusedLocationApi.requestLocationUpdates(client, request, this);
         isRequesting = true;
     }
